Refactor: Consolidate MainActivity logic into BaseActivity and fix intent handling
This commit introduces a `BaseActivity` to share common logic between different `MainActivity` implementations. It also corrects the order of subject and text when sharing a note and updates how note details are extracted from an intent.

**Key Changes:**

*   **`app/src/androidGooglePlay/kotlin/com/hobit/synapse/MainActivity.kt`:**
    *   Now extends `BaseActivity`.
    *   Removed common code that was moved to `BaseActivity`.
    *   Removed unused imports.
*   **`app/src/androidFossReliant/kotlin/com/hobit/synapse/MainActivity.kt`:**
    *   Now extends `BaseActivity`.
    *   Removed common code that was moved to `BaseActivity`.
*   **`core/ui/src/androidMain/kotlin/com/mshdabiola/ui/RealLogics.kt`:**
    *   Swapped `notePad.detail` and `notePad.title` when setting the subject and text for the share intent to ensure correct mapping.
*   **`app/src/androidMain/kotlin/com/hobit/synapse/BaseActivity.kt`:**
    *   Created new `BaseActivity` extending `ComponentActivity`.
    *   Moved common `MainActivity` logic (splash screen, UI state handling, edge-to-edge setup, theme handling, NavHostController setup, and `SynApp` composable invocation) into this class.
    *   Updated the `getNote()` method to correctly extract the title by combining `Intent.EXTRA_TITLE` and `Intent.EXTRA_SUBJECT`, and the note content from `Intent.EXTRA_TEXT`.
    *   Added logging in `getNote()` to print the extracted title, subject, and clip data size for debugging.
